refactor(songs): rename the loop component
