+++ lib/songs/analyzer/analyzer_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:musbx/songs/analyzer/chords_display.dart';
 import 'package:musbx/songs/analyzer/waveform_widget.dart';
-import 'package:musbx/songs/looper/loop_slider.dart';
+import 'package:musbx/songs/loop/loop_slider.dart';
 import 'package:musbx/songs/player/song_player.dart';
 import 'package:musbx/songs/player/songs.dart';
 import 'package:musbx/widgets/flat_card.dart';
